% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colour.R
\name{add_colour}
\alias{add_colour}
\title{Function to add colour or fill to the plot}
\usage{
add_colour(p, color_dim, type = c("colour", "fill"), discrete_cutoff = 9)
}
\arguments{
\item{p}{Plot object created by `p <- ggplot(...)` etc}

\item{color_dim}{Which dimension has been set as the colour aesthetic in `ggplot()`}

\item{type}{Whether it's a fill or colour aesthetic}

\item{discrete_cutoff}{Cutoff below which unique values are assumed to be categorical rather than continuous}
}
\value{
A ggplot layer that will be added to the `p` plot object
}
\description{
This is a helper function used by `theme_cau()`. You shouldn't need to call it directly.
}
