% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colours.R
\name{get_colours}
\alias{get_colours}
\title{Generate a colour palette from the CAU theme}
\usage{
get_colours(n_vals)
}
\arguments{
\item{n_vals}{Number of colours you want in the palette}
}
\value{
A vector of hex colours, of length `n_vals`
}
\description{
Generate a colour palette from the CAU theme
}
